Issue #2625: reorganized internal tests
