+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/CommitValidationTest.java
@@ -17,7 +17,7 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ////////////////////////////////////////////////////////////////////////////////
 
-package com.puppycrawl.tools.checkstyle;
+package com.puppycrawl.tools.checkstyle.internal;
 
 import static org.junit.Assert.assertTrue;
